hri/requirements.txt
openai-whisper
sounddevice
scipy
numpy
pygame
mediapipe
opencv-python
piper-tts
asyncio
aiohttp
